+++ chatette/units/intent/__init__.py
@@ -1,0 +1,3 @@
+from .definition import (IntentDefinition, )
+from .example import (IntentExample, )
+from .rule_content import (IntentRuleContent, )