+++ chatette/units/slot/__init__.py
@@ -1,0 +1,2 @@
+from .definition import (SlotDefinition, )
+from .rule_content import (DummySlotValRuleContent, SlotRuleContent)