+++ chatette/units/word/__init__.py
@@ -1,0 +1,2 @@
+from .group_rule_content import (GroupWordRuleContent, )
+from .rule_content import (WordRuleContent, )